refactor: extract tree-based configuration UI into a reusable TreeEditor component
