+++ src/components/index.js
@@ -6,3 +6,4 @@ export { default as PluginSkeleton } from './skeleton/plugin-skeleton';
 export { default as ThemeSkeleton } from './skeleton/theme-skeleton';
 export { default as StepWrapper } from './step-wrapper';
 export { default as KeyValueEditor } from './key-value-editor';
+export { default as TreeEditor } from './tree-editor';
